Update Lab029_Task 1.py

+++ Nov/ex_09112024_Literals_Variables/Lab029_Task 1.py
@@ -6,9 +6,9 @@ Mul
 Div
 """
 
-fvalue = int(input("Enter 1st value"))
-svalue = int(input("Enter 2st value"))
-tvalue = int(input("Enter 3st value"))
+fvalue = float(input("Enter 1st value"))
+svalue = float(input("Enter 2st value"))
+tvalue = float(input("Enter 3st value"))
 
 sum = fvalue + svalue + tvalue
 sub = fvalue - svalue - tvalue
